Move struct `expander` into header `kernel/expander.hpp`
