+++ src/kernel/syntactic_environment.cpp
@@ -15,317 +15,14 @@
 */
 
 #include <meevax/kernel/boolean.hpp>
-#include <meevax/kernel/conditional_expand.hpp>
 #include <meevax/kernel/environment.hpp>
+#include <meevax/kernel/expander.hpp>
 #include <meevax/kernel/identity.hpp>
-#include <meevax/kernel/include.hpp>
 #include <meevax/kernel/syntactic_environment.hpp>
 #include <meevax/kernel/transformer.hpp>
 
 namespace meevax::inline kernel
 {
-  #define EXPANDER(NAME)                                                       \
-  auto NAME([[maybe_unused]] syntactic_environment const& expander,            \
-                             object const& form,                               \
-            [[maybe_unused]] object const& bound_variables,                    \
-            [[maybe_unused]] typename syntactic_closure::renamer & rename) -> object
-
-  EXPANDER(syntactic_environment::expander::quote)
-  {
-    return form;
-  }
-
-  EXPANDER(syntactic_environment::expander::quote_syntax)
-  {
-    return form;
-  }
-
-  EXPANDER(syntactic_environment::expander::call)
-  {
-    return cons(expander.expand(car(form),
-                                bound_variables,
-                                rename),
-                operand(expander,
-                        cdr(form),
-                        bound_variables,
-                        rename));
-  }
-
-  EXPANDER(syntactic_environment::expander::operand)
-  {
-    if (form.is<pair>())
-    {
-      return cons(expander.expand(car(form),
-                                  bound_variables,
-                                  rename),
-                  operand(expander,
-                          cdr(form),
-                          bound_variables,
-                          rename));
-    }
-    else
-    {
-      return expander.expand(form, bound_variables, rename);
-    }
-  }
-
-  EXPANDER(syntactic_environment::expander::lambda)
-  {
-    auto scoped_rename = rename;
-
-    let const& formals = scoped_rename(cadr(form), bound_variables);
-
-    return cons(rename(car(form)) /* lambda */,
-                cons(formals,
-                     body(expander,
-                          cddr(form),
-                          cons(formals, bound_variables),
-                          scoped_rename)));
-  }
-
-  EXPANDER(syntactic_environment::expander::body)
-  {
-    if (auto [reversed_binding_specs,
-              sequence,
-              current_environment] = expander.sweep(form,
-                                                    form,
-                                                    bound_variables,
-                                                    make<syntactic_environment>(bound_variables, expander.second),
-                                                    rename);
-        reversed_binding_specs)
-    {
-      /*
-         (letrec* <binding specs> <sequence>)
-
-             => ((lambda <variables> <assignments> <sequence>)
-                 <dummy 1> ... <dummy n>)
-
-         where <binding specs> = ((<variable 1> <initial 1>) ...
-                                  (<variable n> <initial n>))
-      */
-      let & formals = caar(current_environment);
-
-      for (let const& binding_spec : reversed_binding_specs)
-      {
-        if (not car(binding_spec).is<macro>()) // The binding-spec is not an internal syntax definition.
-        {
-          sequence = cons(cons(corename("set!"), binding_spec), sequence);
-        }
-      }
-
-      for (let & formal : formals)
-      {
-        if (formal.is<macro>()) // is internal-sytnax-definition
-        {
-          cdr(formal) = make<transformer>(environment().execute(current_environment.template as<syntactic_environment>().compile(cdr(formal) /* <transformer spec> */)),
-                                          current_environment);
-        }
-      }
-
-      return expander.expand(list(cons(cons(corename("lambda"),
-                                            formals,
-                                            sequence),
-                                       make_list(length(formals), unit))),
-                             bound_variables,
-                             rename);
-    }
-    else if (sequence.template is<pair>())
-    {
-      return cons(expander.expand(car(sequence),
-                                  bound_variables,
-                                  rename),
-                  body(expander,
-                       cdr(sequence),
-                       bound_variables,
-                       rename));
-    }
-    else
-    {
-      return expander.expand(sequence, bound_variables, rename);
-    }
-  }
-
-  EXPANDER(syntactic_environment::expander::conditional)
-  {
-    return cons(rename(car(form)),
-                operand(expander,
-                        cdr(form),
-                        bound_variables,
-                        rename));
-  }
-
-  EXPANDER(syntactic_environment::expander::set)
-  {
-    return cons(rename(car(form)),
-                operand(expander,
-                        cdr(form),
-                        bound_variables,
-                        rename));
-  }
-
-  EXPANDER(syntactic_environment::expander::include)
-  {
-    return expander.expand(cons(corename("begin"),
-                                meevax::include(cadr(form))),
-                           bound_variables,
-                           rename);
-  }
-
-  EXPANDER(syntactic_environment::expander::include_case_insensitive)
-  {
-    return expander.expand(cons(corename("begin"),
-                                meevax::include(cadr(form), false)),
-                           bound_variables,
-                           rename);
-  }
-
-  EXPANDER(syntactic_environment::expander::conditional_expand)
-  {
-    return expander.expand(cons(corename("begin"),
-                                meevax::conditional_expand(cdr(form))),
-                           bound_variables,
-                           rename);
-  }
-
-  EXPANDER(syntactic_environment::expander::letrec)
-  {
-    let const extended_bound_variables = cons(rename(map(car, cadr(form)), bound_variables),
-                                              bound_variables);
-
-    return cons(car(form),
-                map([&](let const& binding)
-                    {
-                      return list(car(binding),
-                                  expander.expand(cadr(binding),
-                                                  extended_bound_variables,
-                                                  rename));
-                    },
-                    cadr(form)),
-                body(expander,
-                     cddr(form),
-                     extended_bound_variables,
-                     rename));
-  }
-
-  EXPANDER(syntactic_environment::expander::sequence)
-  {
-    if (form.is<pair>())
-    {
-      return cons(expander.expand(car(form),
-                                  bound_variables,
-                                  rename),
-                  sequence(expander,
-                           cdr(form),
-                           bound_variables,
-                           rename));
-    }
-    else
-    {
-      return expander.expand(form, bound_variables, rename);
-    }
-  }
-
-  EXPANDER(syntactic_environment::expander::let_syntax)
-  {
-    let const current_environment = make<syntactic_environment>(bound_variables, expander.second);
-
-    auto formal = [&](let const& syntax_spec)
-    {
-      return make<macro>(car(syntax_spec) /* keyword */,
-                         make<transformer>(environment().execute(current_environment.as<syntactic_environment>().compile(cadr(syntax_spec) /* transformer spec */)),
-                                           current_environment));
-    };
-
-    let const formals = map(formal, cadr(form));
-
-    return expander.expand(list(cons(corename("lambda"),
-                                     formals,
-                                     cddr(form) /* body */)),
-                           bound_variables,
-                           rename);
-  }
-
-  EXPANDER(syntactic_environment::expander::letrec_syntax)
-  {
-    let current_environment = make<syntactic_environment>(bound_variables, expander.second);
-
-    auto formal = [&](let const& syntax_spec)
-    {
-      return make<macro>(car(syntax_spec) /* keyword */,
-                         make<transformer>(environment().execute(current_environment.as<syntactic_environment>().compile(cadr(syntax_spec) /* transformer spec */)),
-                                           current_environment));
-    };
-
-    let const formals = map(formal, cadr(form));
-
-    current_environment.as<syntactic_environment>().first = cons(formals, bound_variables);
-
-    return expander.expand(list(cons(corename("lambda"),
-                                     formals,
-                                     cddr(form) /* body */)),
-                           bound_variables,
-                           rename);
-  }
-
-  EXPANDER(syntactic_environment::expander::define)
-  {
-    if (cadr(form).is<pair>()) // (define (<variable> . <formals>) <body>)
-    {
-      return list(rename(car(form)),
-                  caadr(form) /* variable */,
-                  expander.expand(cons(corename("lambda"),
-                                       cdadr(form) /* formals */,
-                                       cddr(form) /* body */),
-                                  bound_variables,
-                                  rename));
-    }
-    else // (define <variable> <expression>)
-    {
-      return cons(rename(car(form)),
-                  cadr(form),
-                  cddr(form) ? list(expander.expand(caddr(form),
-                                                          bound_variables,
-                                                          rename))
-                                   : unit);
-    }
-  }
-
-  EXPANDER(syntactic_environment::expander::define_syntax)
-  {
-    return list(rename(car(form)),
-                cadr(form),
-                expander.expand(caddr(form),
-                                bound_variables,
-                                rename));
-  }
-
-  EXPANDER(syntactic_environment::expander::call_with_current_continuation)
-  {
-    return cons(rename(car(form)),
-                operand(expander,
-                        cdr(form),
-                        bound_variables,
-                        rename));
-  }
-
-  EXPANDER(syntactic_environment::expander::current)
-  {
-    return cons(rename(car(form)),
-                operand(expander,
-                        cdr(form),
-                        bound_variables,
-                        rename));
-  }
-
-  EXPANDER(syntactic_environment::expander::install)
-  {
-    return cons(rename(car(form)),
-                operand(expander,
-                        cdr(form),
-                        bound_variables,
-                        rename));
-  }
-
   #define GENERATOR(NAME)                                                      \
   auto NAME([[maybe_unused]] syntactic_environment & generator,                \
             [[maybe_unused]] object const& form,                               \
@@ -861,7 +558,7 @@ namespace meevax::inline kernel
                                rename,
                                formals,
                                cons(list(variable,
-                                         cons(corename("lambda"),
+                                         cons(syntactic_environment::corename("lambda"),
                                               cdadar(sequence), // <formals>
                                               cddar(sequence))),
                                     reversed_binding_specs));
